Move results to new module
